Revert "GameSpeed type - int to double"
This reverts commit 9340844ba5422dcb2a3eb0119cd01128c8bb74ce.

+++ OgameData/OGame.cs
@@ -8,7 +8,7 @@ namespace OgameData
     public class OGame
     {
         [JsonProperty("GameSpeed")]
-        public static double GameSpeed = 1;
+        public static int GameSpeed = 1;
 
         [JsonProperty("Researches")]
         public static Dictionary<Item, Research> Researches = new()
